\name{adja4three}
\alias{adja4three}
\title{
Adjacency matrices of DAGs having three nodes
}
\description{
Returns the list of the 25 adjacency matrices associated to DAGs
comprising three nodes. The first character of the name components
gives the number of arcs in the DAG.
}
\usage{
adja4three(nona=LETTERS[1:3])
}
\arguments{
\item{nona}{ The three node names.}
}
\details{
Poor filling...
}
\value{
a named list having 25 components, each being a 3x3 matrix.
}
