\name{bn2nbn}
\alias{bn2nbn}
\title{
transforms a /bn/ of /bnlearn/ package to a /nbn/
}
\description{
returns a \code{nbn} object from a DAG (\code{bn} object) of
/bnlearn/ package. O and 1 coefficients are introduced...
}
\usage{
bn2nbn(bn)
}
\arguments{
\item{bn}{ The object to be transformed.}
}
\value{
A list following the \code{nbn} specification
}
