% 
% created by /documair/ at 14_09_16@11:37 
% 
\name{complete8rbmn0} 
% 
\alias{complete8rbmn0} 
% 
\title{completes the general constant with further components} 
% 
\usage{ 
  complete8rbmn0(rbmn0) 
} 
% 
\examples{ 
  rbmn0 <- complete8rbmn0(rbmn0); 
} 
% 
\details{ 
   
  More \samp{nbn}, \samp{adja}, \samp{mn}, \samp{gema} and 
  \samp{arc} objects are generated using available functions of 
  \pkg{rbmn} package. Preexisting ones will be silently replaced. 
   
} 
% 
\description{ 
   
  From the initial constant \samp{rbmn0}, returns it 
  with more components, keeping the initial ones. 
   
} 
% 
\value{ 
   
  The completed constant 
   
} 
% 
 
\arguments{ 
    \item{rbmn0}{ The initial constant to be completed.} 
} 
 
% 
\keyword{complete} 
\keyword{rbmn} 
