% 
% created by /documair/ at 14_09_16@11:37 
% 
\name{nbn2pam} 
% 
\alias{nbn2pam} 
% 
\title{gets the /pam/ associate to a /nbn/} 
% 
\usage{ 
  nbn2pam(nbn) 
} 
% 
\description{ 
   
  returns a \samp{pam} object of /g4n/ 
  package with the parentship of \samp{nbn}. 
   
} 
% 
\value{ 
   
  A \samp{pam} object 
   
} 
% 
 
\arguments{ 
    \item{nbn}{ The initial \samp{nbn} object.} 
} 
 
% 
\keyword{normal--Bayesian--network} 
\keyword{pam} 
