\name{reverse8chain}
\alias{reverse8chain}
\title{
reverses the nodes of a chain
}
\description{
returns the chain obtained after reversing its node order
}
\usage{
reverse8chain(chain)
}
\examples{
 print8chain(rbmn0chain.02);
 print8chain(reverse8chain(rbmn0chain.02));
}
\arguments{
\item{chain}{ The chain object to consider.}
}
\value{
The resulting chain
}
