% 
% created by /documair/ at 14_09_16@11:37 
% 
\name{adja4three} 
% 
\alias{adja4three} 
% 
\title{Adjacency matrices of DAGs having three nodes} 
% 
\usage{ 
  adja4three(nona=LETTERS[1:3]) 
} 
% 
\details{ 
   
  Poor filling... 
   
} 
% 
\description{ 
   
  Returns the list of the 25 adjacency matrices associated 
  to DAGs comprising three nodes. The first character 
  of the name components gives the number of arcs in the DAG. 
   
} 
% 
\value{ 
   
  a named list having 25 components, each being a 3x3 matrix. 
   
} 
% 
 
\arguments{ 
    \item{nona}{ The three node names.} 
} 
 
% 
\keyword{adjacency--matrix} 
