% 
% created by /documair/ at 14_09_16@11:37 
% 
\name{bn2nbn} 
% 
\alias{bn2nbn} 
% 
\title{transforms a /bn/ of /bnlearn/ package to a /nbn/} 
% 
\usage{ 
  bn2nbn(bn) 
} 
% 
\description{ 
   
  returns a \samp{nbn} object 
  from a DAG (\samp{bn} object) of /bnlearn/ 
  package. O and 1 coefficients are introduced... 
   
} 
% 
\value{ 
   
  A list following the \samp{nbn} specification 
   
} 
% 
 
\arguments{ 
    \item{bn}{ The object to be transformed.} 
} 
 
% 
\keyword{bnlearn--package} 
\keyword{normal--Bayesian--network} 
