% 
% created by /documair/ at 14_09_16@11:37 
% 
\name{bnfit2nbn} 
% 
\alias{bnfit2nbn} 
% 
\title{transforms a /bn.fit/ of /bnlearn/ package to a /nbn/} 
% 
\usage{ 
  bnfit2nbn(bn.fit) 
} 
% 
\details{ 
   
  If \samp{bn.fit} is not pertinent, a fatal error 
  is issued. 
   
} 
% 
\description{ 
   
  returns a \samp{nbn} object 
  from a Gaussian \samp{bn.fit} object of /bnlearn/ 
  package. 
   
} 
% 
\value{ 
   
  A list following the \samp{nbn} specification 
   
} 
% 
 
\arguments{ 
    \item{bn.fit}{ The object to be transformed.} 
} 
 
% 
\keyword{bnlearn--package} 
\keyword{normal--Bayesian--network} 
